Moved teardown_db_session() to the app_factory
